Time: 2 ms (42.60%), Space: 5.9 MB (56.03%) - LeetHub

+++ 0556-next-greater-element-iii/0556-next-greater-element-iii.cpp
@@ -31,10 +31,10 @@ public:
         }
         swap(nums[pivotIndex],nums[currIndex]);
         sort(nums.begin()+pivotIndex+1,nums.end());
-        res = std::stoll(nums);
+        res = stoll(nums);
         if (res > INT_MAX) {
             return -1;
         }
-        return static_cast<int>(res);
+        return res;
     }
 };
